% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventsObserveR.R
\name{eventsObserverOutput}
\alias{eventsObserverOutput}
\title{eventsObserverOutput}
\usage{
eventsObserverOutput(outputId, width = "100\%", height = "400px")
}
\arguments{
\item{outputId}{output variable to read from}

\item{width}{width of eventsObserveR}

\item{height}{height of eventsObserveR}
}
\description{
Use \code{eventsObserverOutput()} to create a UI element, and \code{renderEventsObserver()} to render the eventsObserveR widget.
}

