% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{eventsObserverOutput}
\alias{eventsObserverOutput}
\alias{renderEventsObserver}
\title{Wrapper functions for using \pkg{leaflet} in \pkg{shiny}}
\usage{
eventsObserverOutput(outputId, width = "100\%", height = "400px")

renderEventsObserver(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{the width and height of the eventsObserver (see
\code{\link[htmlwidgets]{shinyWidgetOutput}})}

\item{expr}{An expression that generates an HTML widget}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Use \code{eventsObserverOutput()} to create a UI element, and \code{renderEventsObserver()}
to render the map widget.
}
\examples{
\dontrun{library(eventsObserveR)
library(shiny)
app = shinyApp(
  ui = fluidPage(eventsObserverOutput('eObserver')),
  server = function(input, output) {
    output$eObserver = renderEventsObserver(eventsObserveR(events = sample_events_data,
place.key = "station"))
  }
)

if (interactive()) print(app)}
}

