% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventsObserveR.R
\name{renderEventsObserver}
\alias{renderEventsObserver}
\title{renderEventsObserver}
\usage{
renderEventsObserver(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{Call to eventsObserveR that generates a eventsObserveR class object that can be rendered client-side using eventsObserveROutput}

\item{env}{The environment in which to evaluate expr.}

\item{quoted}{Is expr a quoted expression (with quote())? This is useful if you want to save an expression in a variable.}
}
\description{
Use \code{eventsObserverOutput()} to create a UI element, and \code{renderEventsObserver()} to render the eventsObserveR widget.
}

