% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_events_data}
\alias{sample_events_data}
\title{Sample events data}
\format{A data frame with 6621 rows and 6 variables:
\describe{
  \item{station}{station, unique identifier for each observation used which must be set as place.key.}
  \item{time}{time, when the event was observed, in as.POSIXct format.}
  \item{event_type}{species, species identification, i.e. cougar, kitten, stick insect.}
  \item{event_type_id}{species, unique id for each species ranging from 0-36}
  \item{radius}{radius, size of event when displayed.}
  \item{color}{color, color for the species type}
  \item{place_id}{placed_id, id of the location (place) where the event was observed, corresponding to the id in `places`}
  \item{title}{title, text to show in the tooltip for each event.}
}}
\usage{
sample_events_data
}
\description{
A sample set of observed events, one might imagine these are sightings of different species of animals at specific locations (motion-activated cameras) - identified by the `station` column. Note that this is a randomised data set and is not indicative of the true migratory behaviour of e.g. meerkats, cougars and polar bears.
}
\keyword{datasets}
