% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_locations_data}
\alias{sample_locations_data}
\title{Sample locations data}
\format{A data frame with 99 rows and 6 variables:
\describe{
  \item{id}{id, Unique id for each location, ranging from 0 - 98. Note that these ids correspond to the `place_id` in `sample_events_data`}
  \item{x}{x, x coordinate of each locations}
  \item{y}{y, y coordinate of each locations}
  \item{color}{color, color of each location}
  \item{radius}{radius, radiues of each location}
  \item{title}{title, tooltip text displayed when hovering over the location}
}}
\usage{
sample_locations_data
}
\description{
The locations of motion-activated camera traps that recorded the events from `sample_events_data`, with (x,y) coordinates. Note that for data sensitivity reasons the coordinates are not required to be valid longitude/latitude pairs.
}
\keyword{datasets}
