% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventsObserveR.R
\name{eventsObserveR}
\alias{eventsObserveR}
\title{eventsObserver}
\usage{
eventsObserveR(events, places = NULL, place.key = "place",
  periodsPerSecond = 24, period = 86400, previousPeriodDuration = 86400,
  size = list(view.width = 600, view.height = 600, interface.width = 1024,
  interface.height = 786, horizontal.margin = 100, vertical.margin = 100),
  place.radius = NULL, event.radius = 5, place.color = "lavenderblush",
  event.color = "red", legend = NULL, legend.columns = 1)
}
\arguments{
\item{events}{A data.frame with events, needs at least place.key, time
\itemize{
 \item{"place"}{ : unique "place" id, does not need to be numeric}
 \item{"time"}{ : integer time of event, cannot be as.POSIXct}
 \item{"title"}{ : tooltip of the event}
 \item{...}{}
 }}

\item{places}{An optional data.frame for place locations, if NULL places will be set to fill outline a circle that fills the available space. Default to NULL
\itemize{
 \item{"place"}{ : unique "place" id, does not need to be numeric}
 \item{"time"}{ : integer time of event, cannot be as.POSIXct}
 \item{"title"}{ : tooltip of the event}
 \item{...}{}
 }}

\item{place.key}{Name of column containing place.key in the events data.frame (defaults to place).}

\item{periodsPerSecond}{Equivalent to number of "frames per second when playing" the eventsObserver animation. Default 24.}

\item{period}{Period within which events must occur to be displayed as filled dots. Default to 86400 seconds.}

\item{previousPeriodDuration}{Period within which events occuring prior to "period" will be included in the visualisation and displayed as empty dots. Default to 86400 seconds.}

\item{size}{Optional list of named arguments:
\itemize{
 \item{"place"}{ : unique "place" id, does not need to be numeric}
 \item{"time"}{ : integer time of event, cannot be as.POSIXct}
 \item{"title"}{ : tooltip of the event}
 \item{...}{}
 }}
}
\description{
\code{eventsObserveR} creates an playable visualisation of the distribution of event observations over a range of defined places within a variable time period.
}

