% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eventsObserveR.R
\name{eventsObserveR}
\alias{eventsObserveR}
\title{eventsObserver}
\usage{
eventsObserveR(events, places = NULL, place.key = "place",
  periodsPerSecond = 24, period = 1, previousPeriodDuration = 1,
  period.unit = "days", previous.period.unit = "days",
  size = list(view.width = 700, view.height = 600, interface.width = 1024,
  interface.height = 786, horizontal.margin = 100, vertical.margin = 100),
  place.radius = NULL, shape.type = "circle", event.radius = 5,
  place.color = "lavenderblush", event.color = "red", legend = FALSE,
  legend.columns = 1)
}
\arguments{
\item{events}{A data.frame with events, needs at least place.key, time and place_id
\itemize{
  \item{place.key}{place.key, unique identifier for each observation used which must be set as place.key.}
  \item{time}{time, when the event was observed, in as.POSIXct format.}
  \item{radius}{radius, size of event when displayed.}
  \item{color}{color, color for the event}
  \item{place_id}{placed_id, id of the location (place) where the event was observed. Must be given if `places` is not NULL.}
  \item{event_type_id}{event_type_id, event type id ranging from 0 - (number_of_event_types - 1). Used by the legend for grouping.}
 }}

\item{places}{An optional data.frame for place locations, if NULL places will be set to fill outline a circle that fills the available space. Default to NULL
\itemize{
  \item{id}{id, Unique id for each location, ranging from 0 - 98. Note that these ids correspond to the `place_id` in `sample_events_data`}
  \item{x}{x, x coordinate of each locations}
  \item{y}{y, y coordinate of each locations}
  \item{color}{color, color of each location}
  \item{radius}{radius, radiues of each location}
  \item{title}{title, tooltip text displayed when hovering over the location}
 }}

\item{place.key}{Name of column containing place.key in the events data.frame (defaults to place).}

\item{periodsPerSecond}{Equivalent to number of "frames per second when playing" the eventsObserver animation. Default 24.}

\item{period}{Period within which events must occur to be displayed as filled dots. Default to 86400 seconds.}

\item{previousPeriodDuration}{Period within which events occuring prior to "period" will be included in the visualisation and displayed as empty dots. Default to 86400 seconds.}

\item{period.unit}{Unit (seconds, hours, days) used to calculate the length of the `period`. Default to "days".}

\item{previous.period.unit}{Unit (seconds, hours, days) used to calculate the length of the `previous.period`. Default to "days".}

\item{size}{Optional list of named arguments:
\itemize{
 \item{"view.width"}{ : width of area containing both the events viewer and the play/pause controls. Default to 700px}
 \item{"view.height"}{ : height of area containing both the events viewer and the play/pause controls. Default to 500px}
 \item{"interface.width"}{ : width of the entire interface, must be larger than view.width. If the legend cannot fit to the right, it will be shown below. Note, however, that the viewer and legend are allowed to rescale themselves dynamically. Default to 1024}
 \item{"interface.height"}{ : height of the entire interface, must be larger than view.height. If the legend cannot fit to the right, it will be shown below. Note, however, that the viewer and legend are allowed to rescale themselves dynamically. Default to 786}
 \item{"horizontal.margin"}{ : margin width placed around the events viewer to prevent events displaying outside the interface. Default to 100px}
 \item{"vertical.margin"}{ : margin height placed around the events viewer to prevent events displaying outside the interface. Default to 100px}
 }}

\item{place.radius}{Radius for circles representing locations for events, used only if `places` is not NULL. Otherwise the radius column in `places` is used.}

\item{shape.type}{Type of shape to use for events, defaults to "circle". Can be path to svg images.}

\item{event.radius}{Radius for events, default to 5. Only used if a column called "radius" in `events` is not given.}

\item{place.color}{Color for places, default to "lavenderblush". Only used if a column called "color" in `places` is not given.}

\item{event.color}{Color of events, only used if a column called "color" is not given in `events`. Default is "red".}

\item{legend}{An optional data.frame for legend labelling events by type. Default to FALSE.
\itemize{
    \item{"description"}{ : unique "place" id, does not need to be numeric}
    \item{"color"}{ : }
}}

\item{legend.columns}{How many columns should the legened entries be split across. Default to 1}
}
\description{
\code{eventsObserveR} creates an playable visualisation of the distribution of event observations over a range of defined places within a variable time period.
}

